Moving the base package to a separate file packet.go
